Trait janus_aggregator_core::query_type::AccumulableQueryType

pub trait AccumulableQueryType: QueryType {
    // Required methods
    fn to_batch_identifier(
        _: &AggregatorTask,
        _: &Self::PartialBatchIdentifier,
        client_timestamp: &Time
    ) -> Result<Self::BatchIdentifier, Error>;
    fn get_conflicting_aggregate_share_jobs<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait, const SEED_SIZE: usize, C, A>(
        tx: &'life0 Transaction<'life1, C>,
        vdaf: &'life2 A,
        task_id: &'life3 TaskId,
        partial_batch_identifier: &'life4 Self::PartialBatchIdentifier,
        report_metadata: &'life5 ReportMetadata
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AggregateShareJob<SEED_SIZE, Self, A>>, Error>> + Send + 'async_trait>>
       where C: 'async_trait + Clock,
             A: 'async_trait + Aggregator<SEED_SIZE, 16> + Send + Sync,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             'life4: 'async_trait,
             'life5: 'async_trait;
    fn get_collection_jobs_including<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait, const SEED_SIZE: usize, C, A>(
        tx: &'life0 Transaction<'life1, C>,
        vdaf: &'life2 A,
        task_id: &'life3 TaskId,
        batch_identifier: &'life4 Self::BatchIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<Vec<CollectionJob<SEED_SIZE, Self, A>>, Error>> + Send + 'async_trait>>
       where C: 'async_trait + Clock,
             A: 'async_trait + Aggregator<SEED_SIZE, 16> + Send + Sync,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             'life4: 'async_trait;
    fn to_batch_interval(
        batch_identifier: &Self::BatchIdentifier
    ) -> Option<&Interval>;
    fn downgrade_batch_identifier(
        batch_identifier: &Self::BatchIdentifier
    ) -> &Self::PartialBatchIdentifier;
    fn upgrade_partial_batch_identifier(
        partial_batch_identifier: &Self::PartialBatchIdentifier
    ) -> Option<&Self::BatchIdentifier>;
    fn default_partial_batch_identifier(
    ) -> Option<&'static Self::PartialBatchIdentifier>;
}

Required Methods

fn to_batch_identifier( _: &AggregatorTask, _: &Self::PartialBatchIdentifier, client_timestamp: &Time ) -> Result<Self::BatchIdentifier, Error>

This method converts various values related to a client report into a batch identifier. The arguments are somewhat arbitrary in the sense they are what “works out” to allow the necessary functionality to be implemented for all query types.

fn get_conflicting_aggregate_share_jobs<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait, const SEED_SIZE: usize, C, A>( tx: &'life0 Transaction<'life1, C>, vdaf: &'life2 A, task_id: &'life3 TaskId, partial_batch_identifier: &'life4 Self::PartialBatchIdentifier, report_metadata: &'life5 ReportMetadata ) -> Pin<Box<dyn Future<Output = Result<Vec<AggregateShareJob<SEED_SIZE, Self, A>>, Error>> + Send + 'async_trait>>where C: 'async_trait + Clock, A: 'async_trait + Aggregator<SEED_SIZE, 16> + Send + Sync, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, 'life4: 'async_trait, 'life5: 'async_trait,

fn get_collection_jobs_including<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait, const SEED_SIZE: usize, C, A>( tx: &'life0 Transaction<'life1, C>, vdaf: &'life2 A, task_id: &'life3 TaskId, batch_identifier: &'life4 Self::BatchIdentifier ) -> Pin<Box<dyn Future<Output = Result<Vec<CollectionJob<SEED_SIZE, Self, A>>, Error>> + Send + 'async_trait>>where C: 'async_trait + Clock, A: 'async_trait + Aggregator<SEED_SIZE, 16> + Send + Sync, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, 'life4: 'async_trait,

Retrieves collection jobs which include the given batch identifier.

fn to_batch_interval( batch_identifier: &Self::BatchIdentifier ) -> Option<&Interval>

Some query types (e.g. TimeInterval) can represent their batch identifiers as an interval. This method extracts the interval from such identifiers, or returns None if the query type does not represent batch identifiers as an interval.

fn downgrade_batch_identifier( batch_identifier: &Self::BatchIdentifier ) -> &Self::PartialBatchIdentifier

Downgrade a batch identifier into a partial batch identifier.

fn upgrade_partial_batch_identifier( partial_batch_identifier: &Self::PartialBatchIdentifier ) -> Option<&Self::BatchIdentifier>

Upgrade a partial batch identifier into a batch identifier, if possible.

fn default_partial_batch_identifier( ) -> Option<&'static Self::PartialBatchIdentifier>

Get the default value of the partial batch identifier, if applicable.