Trait janus_aggregator_core::query_type::CollectableQueryType

pub trait CollectableQueryType: AccumulableQueryType {
    type Iter: Iterator<Item = Self::BatchIdentifier> + Send + Sync;

    fn collection_identifier_for_query<'life0, 'life1, 'life2, 'life3, 'async_trait, C>(
        tx: &'life0 Transaction<'life1, C>,
        task: &'life2 Task,
        query: &'life3 Query<Self>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self::BatchIdentifier>, Error>> + Send + 'async_trait>>
    where
        C: 'async_trait + Clock,
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait;
    fn batch_identifiers_for_collection_identifier(
        _: &Task,
        collection_identifier: &Self::BatchIdentifier
    ) -> Self::Iter;
    fn validate_collection_identifier(
        task: &Task,
        collection_identifier: &Self::BatchIdentifier
    ) -> bool;
    fn count_client_reports<'life0, 'life1, 'life2, 'life3, 'async_trait, C>(
        tx: &'life0 Transaction<'life1, C>,
        task: &'life2 Task,
        collection_identifier: &'life3 Self::BatchIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>
    where
        C: 'async_trait + Clock,
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait;
    fn acknowledge_collection<'life0, 'life1, 'life2, 'life3, 'async_trait, C>(
        tx: &'life0 Transaction<'life1, C>,
        task_id: &'life2 TaskId,
        batch_identifier: &'life3 Self::BatchIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        C: 'async_trait + Clock,
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait;

    fn get_batch_aggregations_for_collection_identifier<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait, const SEED_SIZE: usize, A, C>(
        tx: &'life0 Transaction<'_, C>,
        task: &'life1 Task,
        vdaf: &'life2 A,
        collection_identifier: &'life3 Self::BatchIdentifier,
        aggregation_param: &'life4 A::AggregationParam
    ) -> Pin<Box<dyn Future<Output = Result<Vec<BatchAggregation<SEED_SIZE, Self, A>>, Error>> + Send + 'async_trait>>
    where
        A::AggregationParam: Send + Sync,
        A::AggregateShare: Send + Sync,
        A: 'async_trait + Aggregator<SEED_SIZE, 16> + Send + Sync,
        C: 'async_trait + Clock,
        Self: Send + 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
    { ... }
}

CollectableQueryType represents a query type that can be collected by Janus. This trait extends AccumulableQueryType with additional functionality required for collection.

Required Associated Types

type Iter: Iterator<Item = Self::BatchIdentifier> + Send + Sync

Required Methods

fn collection_identifier_for_query<'life0, 'life1, 'life2, 'life3, 'async_trait, C>(
tx: &'life0 Transaction<'life1, C>,
task: &'life2 Task,
query: &'life3 Query<Self>
) -> Pin<Box<dyn Future<Output = Result<Option<Self::BatchIdentifier>, Error>> + Send + 'async_trait>>where
C: 'async_trait + Clock,
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,

Retrieves the batch identifier for a given query.

fn batch_identifiers_for_collection_identifier(
_: &Task,
collection_identifier: &Self::BatchIdentifier
) -> Self::Iter

Some query types (e.g. TimeInterval) can receive a batch identifier in collection requests which refers to multiple batches. This method takes a batch identifier received in a collection request and provides an iterator over the individual batches’ identifiers.

fn validate_collection_identifier(
task: &Task,
collection_identifier: &Self::BatchIdentifier
) -> bool

Validates a collection identifier, per the boundary checks in https://www.ietf.org/archive/id/draft-ietf-ppm-dap-02.html#section-4.5.6.

fn count_client_reports<'life0, 'life1, 'life2, 'life3, 'async_trait, C>(
tx: &'life0 Transaction<'life1, C>,
task: &'life2 Task,
collection_identifier: &'life3 Self::BatchIdentifier
) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>where
C: 'async_trait + Clock,
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,

Returns the number of client reports included in the given collection identifier, whether they have been aggregated or not.

fn acknowledge_collection<'life0, 'life1, 'life2, 'life3, 'async_trait, C>(
tx: &'life0 Transaction<'life1, C>,
task_id: &'life2 TaskId,
batch_identifier: &'life3 Self::BatchIdentifier
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
C: 'async_trait + Clock,
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,

Acknowledges that a collection attempt has been made, allowing any query-type specific updates to be made. For exmaple, a task using fixed-size queries might remove the given batch to be removed from the list of batches ready to be returned by a current-batch query.

Provided Methods

fn get_batch_aggregations_for_collection_identifier<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait, const SEED_SIZE: usize, A, C>(
tx: &'life0 Transaction<'_, C>,
task: &'life1 Task,
vdaf: &'life2 A,
collection_identifier: &'life3 Self::BatchIdentifier,
aggregation_param: &'life4 A::AggregationParam
) -> Pin<Box<dyn Future<Output = Result<Vec<BatchAggregation<SEED_SIZE, Self, A>>, Error>> + Send + 'async_trait>>where
A::AggregationParam: Send + Sync,
A::AggregateShare: Send + Sync,
A: 'async_trait + Aggregator<SEED_SIZE, 16> + Send + Sync,
C: 'async_trait + Clock,
Self: Send + 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,

Retrieves batch aggregations corresponding to all batches identified by the given collection identifier.